requirements.txt
Django
django-enumfields
djangorestframework
django-filter
markdown
django-rest-swagger
